Move highestRiskOfHarm function to helpers
