+++ app/data/session-data-defaults.js
@@ -1,5 +1,7 @@
+const path = require('path')
 const cases = require('./cases')
 const fakeCases = require('./fake-cases')
+const helpers = require(path.join(__dirname, '../../lib/helpers.js'))
 
 const contactHistoryDefaults = (map, su) => {
   if (su.contactHistory) {
@@ -8,25 +10,8 @@ const contactHistoryDefaults = (map, su) => {
   return map
 }
 
-const highestRiskOfSeriousHarm = (su) => {
-  const riskValue = {
-    'Low': 0,
-    'Medium': 1,
-    'High': 2,
-    'Very high': 3
-  }
-
-  const compareRisk = (a, b) => {
-    if (riskValue[a.text] < riskValue[b.text]) { return -1 }
-    if (riskValue[a.text] > riskValue[b.text]) { return 1 }
-    return 0
-  }
-
-  su.riskOfSeriousHarmLevel = su.riskOfHarm.map(r => r.inCommunity).sort(compareRisk).pop()
-}
-
 const sessionCases = cases.concat(fakeCases(10))
-sessionCases.forEach(su => highestRiskOfSeriousHarm(su))
+sessionCases.forEach(su => helpers.highestRiskOfSeriousHarm(su))
 
 module.exports = {
   features: {
