+++ lib/helpers.js
@@ -230,6 +230,23 @@ exports.getFlag = (flagId) => {
   return riskFlags.flags.find(flag => flag.id === flagId)
 }
 
+exports.highestRiskOfSeriousHarm = (suCase) => {
+  const riskValue = {
+    'Low': 0,
+    'Medium': 1,
+    'High': 2,
+    'Very high': 3
+  }
+
+  const compareRisk = (a, b) => {
+    if (riskValue[a.text] < riskValue[b.text]) { return -1 }
+    if (riskValue[a.text] > riskValue[b.text]) { return 1 }
+    return 0
+  }
+
+  suCase.riskOfSeriousHarmLevel = suCase.riskOfHarm.map(r => r.inCommunity).sort(compareRisk).pop()
+}
+
 exports.addHelpers = function (env) {
   env.addGlobal('progressPercentage', exports.progressPercentage)
